+++ R/plot_geno.R
@@ -1,6 +1,0 @@
-#' Plot genotypes for one individual
-#'
-#' Plot imputed genotypes across the genome for a single individual
-#'
-#' @param geno Imputed genotypes, as a list of matrices
-#' @param ind  Single numeric index or character string with the genotypes