+++ R/plot_scan1.R
@@ -120,8 +120,8 @@ plot_scan1 <-
                  xlim=NULL, ylim=NULL, xaxs="i", yaxs="i",
                  main="", mgp.x=c(2.6, 0.5, 0), mgp.y=c(2.6, 0.5, 0),
                  mgp=NULL, las=1,
-                 hlines=NULL, hlines.col="white", hlines.lwd=1, hlines.lty=1,
-                 vlines=NULL, vlines.col="white", vlines.lwd=1, vlines.lty=1,
+                 hlines=NULL, hlines_col="white", hlines_lwd=1, hlines_lty=1,
+                 vlines=NULL, vlines_col="white", vlines_lwd=1, vlines_lty=1,
                  ...)
         {
             dots <- list(...)
@@ -197,11 +197,11 @@ plot_scan1 <-
                 # grid lines
                 if(onechr && !(length(vlines)==1 && is.na(vlines))) { # if vlines==NA (or mult chr), skip lines
                     if(is.null(vlines)) vlines <- pretty(xlim)
-                    abline(v=vlines, col=vlines.col, lwd=vlines.lwd, lty=vlines.lty)
+                    abline(v=vlines, col=vlines_col, lwd=vlines_lwd, lty=vlines_lty)
                 }
                 if(!(length(hlines)==1 && is.na(hlines))) { # if hlines==NA, skip lines
                     if(is.null(hlines)) hlines <- pretty(ylim)
-                    abline(h=hlines, col=hlines.col, lwd=hlines.lwd, lty=hlines.lty)
+                    abline(h=hlines, col=hlines_col, lwd=hlines_lwd, lty=hlines_lty)
                 }
             }